Crate fedimint_walletv2_common

Modules

config

endpoint_constants

Structs

DepositRange

FederationWallet

TxInfo

UnknownWalletInputVariantError

UnknownWalletOutputVariantError

WalletCommonInit

WalletInputV0

WalletModuleTypes

WalletOutputOutcome

WalletOutputV0

Enums

StandardScript

WalletConsensusItem

WalletInput

WalletInputError

WalletOutput

WalletOutputError

Constants

KIND

MODULE_CONSENSUS_VERSION

Functions

descriptor

is_potential_receive

Returns true if the script pubkey potentially belongs to the federation. This uses a probabilistic filter - only ~1/65536 of P2WSH scripts pass.

sleep_duration

Returns a sleep duration of 1 second in test environments or 60 seconds in production. Used for polling intervals where faster feedback is needed during testing.

tweak_public_key